+++ phpBB/includes/functions_convert.php
@@ -1952,7 +1952,7 @@ function update_dynamic_config()
 		FROM ' . ATTACHMENTS_TABLE . '
 		WHERE is_orphan = 0';
 	$result = $db->sql_query($sql);
-	set_config('upload_dir_size', (int) $db->sql_fetchfield('stat'), true);
+	set_config('upload_dir_size', (float) $db->sql_fetchfield('stat'), true);
 	$db->sql_freeresult($result);
 
 	/**
